Extracted generating project name and icon to a separate class
